embedding/scripts/requirements.txt
chromadb>=1.0.15
openai>=1.40.0
pyyaml>=6.0.1
glob2>=0.7
